goals into master

+++ AAI-Final-Assignment-WinForms/Entities/BaseGameEntity.cs
@@ -12,7 +12,7 @@ namespace AAI_Final_Assignment_WinForms.Entities
         public Vector2D Pos { get; set; }
 
         // scale size of entity
-        public double Scale { get; set; }
+        public float Scale { get; set; }
 
         // world that contains the entity
         public GameWorld World { get; set; }
@@ -24,7 +24,7 @@ namespace AAI_Final_Assignment_WinForms.Entities
         public int Id { get; set; }
 
         // bounding radius of a object
-        public double Radius { get; set; }
+        public float Radius { get; set; }
 
 
         // texture setup 
@@ -33,8 +33,8 @@ namespace AAI_Final_Assignment_WinForms.Entities
         public Bitmap? Texture { get; set; }
 
 
-        protected BaseGameEntity(Vector2D pos, GameWorld world, double scale, int textureWidth, int textureHeight,
-            double radius)
+        protected BaseGameEntity(Vector2D pos, GameWorld world, float scale, int textureWidth, int textureHeight,
+            float radius)
         {
             Pos = pos;
             World = world;
@@ -57,7 +57,7 @@ namespace AAI_Final_Assignment_WinForms.Entities
         }
 
         // every entity needs a update function
-        public virtual void Update(double timeElapsed)
+        public virtual void Update(float timeElapsed)
         {
         }
